+++ doc/changelog.md
@@ -6,7 +6,9 @@
   - Allow 'build' and/or 'run' configuration to be skipped (#207)
   - Refactored to use 'inspect' instead of 'list' for checking the existence of an image (#230)
   - Refactored ApacheHttpClientDelegate to avoid leaking connections (#232)
-  
+  - Introduced global `portPropertyFile` setting (#90)
+  - Allow the container's host ip to be bound to a maven property and exported
+
 * **0.13.2**
   - "run" directives can be added to the Dockerfile (#191)
   - Support user information in wait URL (#211)